src/Python/requirements.txt
python-dateutil==2.9.0
requests==2.32.5
httpx==0.28.1
yfinance==0.2.65
pandas==2.2.3
pandas-market-calendars==5.1.1


